backend/pyproject.toml
[project]
name = "openneuro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "litellm>=1.81.8",
    "mypy>=1.19.1",
    "numpy>=2.4.2",
    "onnxruntime>=1.24.1",
    "openai>=2.17.0",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.5",
    "sse-starlette>=3.2.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "transformers>=5.1.0",
    "types-requests>=2.32.4.20260107",
    "uvicorn[standard]>=0.30.0",
    "websockets>=16.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.mypy]
[[tool.mypy.overrides]]
module = ["sounddevice", "onnxruntime"]
ignore_missing_imports = true
